requirements.txt
click==8.1.3
filelock==3.12.0
Flask==2.2.3
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
Pillow==9.5.0
sympy==1.11.1
torch==2.0.0
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
